// src/models/IoK8sApiResourceV1alpha2ResourceClaimTemplateList/types.ts
import { IoK8sApiResourceV1alpha2ResourceClaimTemplateModelOptions as IoK8sApiResourceV1alpha2ResourceClaimTemplate } from '../IoK8sApiResourceV1alpha2ResourceClaimTemplate';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiResourceV1alpha2ResourceClaimTemplateListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiResourceV1alpha2ResourceClaimTemplate>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}